refactor(ward): post api
