+++ src/ward/ward.module.ts
@@ -1,11 +1,12 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { Hospital } from 'src/hospital/entities/hospital.entity';
 import { Ward } from './entities/ward.entity';
 import { WardController } from './ward.controller';
 import { WardService } from './ward.service';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Ward])],
+  imports: [TypeOrmModule.forFeature([Ward, Hospital])],
   controllers: [WardController],
   providers: [WardService],
 })
